add an external drive mount monitor
Add a dummy file to the external mount as part of the mounting
process and kick off a process to monitor that file for any
changes. File changes should trigger karen to run a script
that does some checks to establish whether the mount is ok
or if umbrel should be stopped.