check for invalid 'RELEASE' values in update script
This adds two checks for invalid RELEASE values that
can potentially come from the 'update-status.json'
file.

The first checks that RELEASE is not empty. This can
happen if for example the update script was run when

the 'update-status.json' file is in its default
success state and the 'updateTo' value is an empty
string.

The second checks that RELEASE points to a valid URL
to avoid any further error from attempting to do
operations on an invalid curl call.