Draft: Add screen driver
This is a proposal for a screen driver feature for Umbrel. This is the host part. If you think this is a good way to do this, corresponding manager logic will be added.

My proposal would be that apps can request an additional dependency "framebuffer", which gives them access to /dev/fb* and installation causes the manager to execute this trigger, and the dashboard to prompt for a reboot.

This would allow getting apps like https://github.com/lorenzoPrimi/umbrel-crypto-display to work inside docker.

@lukechilds @mayankchhabra What are your thoughts on this?

I'll follow up with PRs to other components if you think this is an acceptable way of adding screen apps.