Append to instead of overwriting resolv.conf
Some networks do not have access to public DNS servers (such as Google
or Cloudflare), and this breaks auto-update for Umbrel in a way that's
very tricky to workaround.

The public DNS's are required for Umbrel users who are using Pi-hole (in
which case their router's DNS may be configured to point to the Umbrel
instance for DNS services, which also breaks updates!).

To manage both of these contingencies, this commit appends the public
DNS servers to the temporary resolv.conf, as well as setting the
failover timeout to 1 second (instead of the default 5 seconds), so that
non-pinhole users will simply use the original default, while Pi-hole
users can just failover gracefully to the public ones.