Fix dev environment
We can't run relaxed CORS for dev because it's too different to production. It will break cookies and therefore the auth server. It will also allow things to work that will never work in production. If we instead proxy the root through the ui dev server we can run everything on the same origin just like it is in production and avoid different CORS behaviour in dev.