Explicitly state the package contents
This is needed because when globally installing a local package with `npm install --global .` it will infer the package contents by excluding files in .gitignore. This means that the ./ui directory in the umbreld project which contains the build output from the ui package will be excluded since it's not tracked in version control.