Ensure app store initialisation retry logic is applied
We had retry logic before but it wasn't working because
`AppStore.update()` never throws it silently fails to be safe to run in
a loop without causing exceptions. Now we have custom logic to only pull
the default app repo and throw on error wrapped in the same retry logic
as before. This is needed as a short term fix for the DHCP race
condition where we try to initialise the App Store before the device has
an IP.