Safer device formatting
Only format the device if it doesn't look like a full bootable umbrelOS
install. This is to protect users who boot from USB or NVMe on Raspberry
Pi. If they temporarily boot from an SDcard image it will try to format
the connected USB/NVMe device to be used as a data drive, wiping all
their data. If the device looks like a bootable umbrelOS install we will
bail from the mount script instead of formatting it. If the user wants
to use this as a data drive they will need to manually format it. They
can format / wipe it in any way, as long as it doesn't look like a
umbrelOS install it will be correctly reformatted by umbrelOS to be used
as a data drive.