Fix HTTPS issue by using APP_PROTOCOL in axios baseURL
This change fixes a compatibility issue with HTTPS when using the Cloudflare Tunnel to expose the Umbrel instance. The issue occurred because axios was configured with a fixed HTTP protocol. By modifying the configuration to use the protocol defined in the APP_PROTOCOL constant, all requests are now correctly made with HTTP or HTTPS as configured, preventing the browser from blocking requests when HTTPS is in use.