Reduce log verbosity
We previously had the log level set to verbose on production. This
change moves all the useful logs out of the verbose log level to the
normal log level and runs the normal log level in production. The only
thing left in the verbose log level is now the HTTP requests which adds
a lot of noise and isn't that useful. Errored out HTTP requests are
still logged in normal mode.