Support fuzzy search in app store
Implement fuzzy search in the app store, with both API and configuration
factored out into a separate file so we can tweak the options globally
or replace the library when we see a need to. Enables use cases such as
searching for "pi hole" and finding Pi-Hole, or searching for "plexx"
and finding Plex.