Animate folder icon on parent hover and optimize the animation
Fixes the following for the animated folder icon:

1) Allows the icon to be animated based on a trigger from a parent div.
We use this so that the icon animates when we hover on the parent that
contains the icon, folder name, type, etc instead of needed to hover
directly on the svg.

2) Refactors the animations to use simpler transforms to increase
performance. No performance loss even when we render 100 icons in our
non-virtualized listing.

3) It also fixes the weird behaviour where the first hover would animate
the overlay icon nicely, but subsequent hovers would no longer show a
nice transition.