Better handling of broken app installs
If an app install is broken we might fail to query it's dependencies. `apps.getDependents` will fail if we can't read the dependencies of any app. Since the ui calls this before doing an uninstall, if any app install is broken in a way that it's dependencies can't be read, no app is able to do an uninstall. We now just ignore any apps that don't return dependencies instead of failing.