Context menu fixes
Disables the browser's default context menu everywhere in the UI
so that we only show a context menu when we've implemented a purposeful,
custom one. It also updates the shadcn components for context-menu,
dropdown-menu, and popover so that we don't get event propagation that
results in showing the context menu from the component that lies
underneath the open menu (including the default browser one).