Handle onboarding language selection edge-case
Adds an explicit check for isOnboarding. Now, the condition 
`if (isOnboarding || !hasJwt)` ensures that if the user is either 
in the onboarding flow OR does not have a JWT, we always take the
local update path (i18next.changeLanguage + window.location.reload)
without calling to the backend.

Technically we don't even need the `!hasJwt` check because we have
routing guards that prevent an unauthenticated user from reaching pages
outside of `/onboarding`, but I've left it in for now because it doesn't
really cost anything to check and it could potentially help in other
edge case scenarios.