Improve thumbnail integration test reliability
The main culprit for the intermittent failing tests seems to be a
timeout issue for background thumbnail generation. When running tests
locally on a powerful machine, the 5s poll we were using for background
thumbnail generation is fine. But in CI this is sometimes not enough to
cover:
- the 1000ms debounce we have in thumbnails.ts that resets on each event
for a path (e.g., multiple update's to a file during copy). We have this
so that we don't generate multiple thumbnails while a file is being
copied/created.
- the actual time it takes to generate the thumbnail via convert/FFmpeg
- general CI resource issues / variability etc.

Other tests that rely on fs events don't seem to intermittently fail
(e.g. recents), but it still might be possible for an fs event to be
missed and result in a test failing.

The general fix is:

1) Extend the time that we wait for thumbnail generation. I've added an
explicit wait period for the thumbnail generation debounce (1500ms)
mostly so it is explicit... But the increased polling time for the
thumbnail generation (now 15s) technically would handle this by itself.

2) Use Vitest's built in `retry` functionality for the background
thumbnail generation tests. For these tests I've just set retries to 5.
If we get a failing test then vitest will retry for a passing test up to
a total of 5 times