Correctly handle long initial backup time
We backup on hourly intervals and return an error notification to the
user if backups have failed for 24 hours. However if the first backup
takes longer than an hour then on the first scheduled backup, the
`lastBackup` is cast to `0` which is intepreted as the unix epoch 1970
and triggers the "no backups for 24 hours" notification.

Now if `lastBackup` is unset we cast `lastBackup` to the time the
backups module started. That when a backup is intially created it can
take up to 24 hours to do it's first backup before triggering the failed
backups error.

Also improved the test to both fail on the error case (backup showing as
failing before 24 hours have passed) and pass on the success case
(backup notification set when failure happens after 24 hours.)