Fix partition expansion on first boot
We used to install `systemd-boot` in Debian Bookworm which includes
`systemd-repart`. Now we're using Trixie we need to install
`systemd-repart` directly as it's own package.

On top of that, the new behaviour fails the partition job if a partition
already exists that can't be resized to exactly match our config. This
wasn't the case in Bookworm. Our system partitions end up as 9.6 GB but
the config specifies at least 10GB. This was causing the job to always
fail (once we actually installed the tool).

We now set min 9GB max 10GB.