Add better handling to shutdown methods
Before this change we would naively wait for backup jobs to finish
during the shutdown process. This is suboptimal because a backup could
take a long time and it's confusing to delay shutdown for such a long
time. Additionally if we're backing up to an unstable USB/Samba share
some weird things can happen where kopia commands can hang for a long
period of time.

We now store a reference to all running kopia processes and send an exit
signal to them at shutdown time. We also don't allow any new kopia
processes to be spawned during the shutdown process.

We also wrap the external storage and network storage shutdown methods in timeouts to avoid hangs there if dealing with temperamental network/usb storage.