Add per-file compression for ZIP downloads
Currently, when downloading multiple-files we stream a ZIP archive to
the client with the default deflate compression behaviour. So no matter
what the file type is we compress it. There's a ton of media file types
that we don't need to do this to because they are already compressed
formats and don't get much (if any) size reduction from compression, but
they incur significant CPU cost when being deflated. So say a user
downloads a folder of tv shows for example, all the shows would be run
through deflate compression and the download would be CPU constrained.

Additionally, due to the way DEFLATE works, highly compressible input compresses much slower due to DEFLATE having to search much harder for recurring patterns. So not only does it waste CPU cycles for no benefit, it also takes significantly longer to compress per input byte, throttling the download speed for the entire zip.

With this PR we now only compress when a file benefits from deflate
compression. So text files like JSON, CSV, logs, and other text-based
formats are compressed, but binary and already-compressed formats like
videos, images, audio, and archives aren't.