Add Umbrel Pro hardware submodule
Initial implementation only supports fan control.

The Umbrel Pro has a single fan that cools both the CPU and all NVMe
SSDs. The EC firmware has its own fan curve for the CPU. We calculate
fan curves for each SSD in umbreld, take the highest value, and set that
as the minimum fan speed in the EC. The EC then uses whichever is higher
- our SSD minimum or its CPU curve.